Import hs-cabal-1.6.0.2 as wip/hs-cabal.
The Haskell Cabal is is the Common Architecture for Building
Applications and Libraries. It is a framework which defines a common
interface for authors to more easily build their applications in a
portable way. The Haskell Cabal is meant to be a part of a larger
infrastructure for distributing, organizing, and cataloging Haskell
Libraries and Tools.

Specifically, the Cabal describes what a Haskell package is, how these
packages interact with the language, and what Haskell implementations
must to do to support packages. The Cabal also specifies some
infrastructure (code) that makes it easy for tool authors to build and
distribute conforming packages.

The Cabal is only one contribution to the larger goal. In particular,
the Cabal says nothing about more global issues such as how authors
decide where in the module name space their library should live; how
users can find a package they want; how orphan packages find new
owners; and so on.