Import gtg-0.1.1 as wip/gtg.
Getting  Things  GNOME!  is an organizer for the GNOME desktop environ-
ment.  GTG focuses on usability and ease of use. Its main objective  is
to  provide  a  simple  and yet flexible organization tool for life and
work.
GTG uses a very handy text edition system for task  creation  and  edi-
tion.   The  task  editor  can automatically recognize metadata such as
tags and subtasks only by the use of a very simple syntax