Psi 0.13-rc4 is here! Please download and test [1].
Final on July 26th if no showstoppers are found.

New in 0.13
- Voice calls (Jingle RTP).
- Basic XMPP URI handler.
- Ability to permanantly trust certificates at connect time.
- Mini command system (Ctrl+7 in chat window).
- Various bugfixes.

If you are building from source (e.g. on Linux) and want voice
calls, you will need to obtain the PsiMedia plugin separately [2].
On Linux, the plugin file is called libgstprovider.so, and must be
put in Psi's $LIBDIR/psi/plugins directory. You'll know it worked
if "About GStreamer" appears in the Help menu.