Import GeoIP-1.4.6 as wip/GeoIP.
GeoIP is a C library that enables the user to find the country that any
IP address or hostname originates from.  It uses a file based database
that is accurate as of March 2002.  This database simply contains IP blocks
as keys, and countries as values.  This database should be more complete and
accurate than using reverse DNS lookups.