Import py25-rdflib-2.4.2 as wip/py-rdflib.
RDFLib is a Python library for working with RDF, a simple yet powerful
language for representing information.

The library contains parsers and serializers for RDF/XML, N3, NTriples,
Turtle, TriX and RDFa . The library presents a Graph interface which can
be backed by any one of a number of Store implementations, including, Memory,
MySQL, Redland, SQLite, Sleepycat, ZODB and SQLObject.