Import hs-event-list-0.0.10 as wip/hs-event-list.
These lists manage events that are associated with times. Times may be
given as difference between successive events or as absolute time
values. Pauses before the first and after the last event are
supported. The underlying data structures are lists of elements of
alternating types, that is [b,a,b,...,a,b] or [a,b,a,...,a,b]. The
data structures can be used to represent MIDI files, OpenSoundControl
message streams, music performances etc.