Import p5-Net-Twitter-Lite-0.08005 as wip/p5-Net-Twitter-Lite.
This module provides a perl interface to the Twitter APIs. It uses the same
API definitions as Net::Twitter, but without the extra bells and whistles and
without the additional dependencies. Same great taste, less filling.

This module is related to, but is not part of the Net::Twitter distribution.
It's API methods and API method documentation are generated from
Net::Twitter's internals. It exists for those who cannot, or prefer not to
install Moose and its dependencies.

You should consider upgrading to Net::Twitter for additional functionality,
finer grained control over features, full backwards compatibility with older
versions of Net::Twitter, and additional error handling options.