Import moonlight-1.0.1 as wip/moonlight.
Moonlight is an open source implementation of Silverlight
(http://silverlight.net), primarily for Linux and other Unix/X11 based
operating systems. In September of 2007, Microsoft and Novell announced a
technical collaboration that includes access to Microsoft's test suites for
Silverlight and the distribution of a Media Pack for Linux users that will
contain licensed media codecs for video and audio.