Import vice-2.1 as wip/vice2.
A newer version of emulators/vice, because its maintainer doesn't
react to mail pointing out newer versions.

VICE is the one and only Versatile Commodore Emulator. It provides
emulation of the Commodore 64, 128, VIC20 and PET 8-bit computers
within a single package. The emulators run as separate programs, but
have the same user interface, share the same settings and support the
same file formats.

All the emulators provide an accurate 6502/6510 emulator, with
emulation of all the opcodes (both documented and undocumented ones)
and accurate timing. Unlike other emulators, VICE aims to be cycle
accurate; it tries to emulate chip timings as precisely as possible
and does so efficiently.

Please do not expect the VIC20, C128 and PET emulators to be as good
as the C64 one, as they are still under construction.

The current version emulates the C64, the C128, the VIC20, the CBM-II
line (B128, B256), and all the PET models.