Import e-uae-0.8.29wip4 as wip/e-uae.
Newer "-WIP4" version than in emulators/e-uae, but nevertheless it hasn't
been updated in nearly 3 years.

This is a fork of UAE, the Ubiquitous Amiga Emulator, with an emulation core
based on WinUAE. It attempts to bring many of the great features of WinUAE
to non-Windows platforms. This version now finally has a name, E-UAE, since
that's what everybody was calling it anyway. The 'E' can stand for anything
you fancy. Experimental, extreme, exciting, egalitarian, eggplant...

E-UAE is open-source software and is made available under the terms of the
GPL. See the COPYING file included with the E-UAE archive for details.

E-UAE is based on the work of dozens of contributors including Bernd Schmidt
(the original author and maintainer of UAE), Bernie Meyer (the author of the
x86 JIT compiler), Toni Wilen (the current maintainer of WinUAE), and many
more.

To make full use of E-UAE you will need access to an image of some version
of the Amiga Kickstart ROM (although UAE does include a ROM emulation which
may work with some old games). The Amiga Kickstart ROMs are copyrighted
material and may not be freely distributed (so don't ask me for a copy). If
you don't have an Amiga from which you can legally make a ROM image, a
selection of ROM images are available for purchase online from Cloanto
(publishers of the Amiga Forever distribution). See
http://www.amigaforever.com/