Import Mesa-7.7 as wip/Mesa.
Finally, the Mesa meta-package !

MesaLib is a 3-D graphics library with an API which is very similar to
that of OpenGL*.  To the extent that Mesa utilizes the OpenGL command syntax
or state machine, it is being used with authorization from Silicon Graphics,
Inc.  However, the author makes no claim that Mesa is in any way a
compatible replacement for OpenGL or associated with Silicon Graphics, Inc.