Import cinelerra-cv-2.1nb20100202 as wip/cinelerra-cv.
Cinelerra is the most advanced non-linear video editor and compositor
for Linux. It is developed by Adam Williams (formerly known as Jack
Crossfire) at Heroine Virtual Ltd..

To know more see the official Cinelerra home page and the SourceForge
project page.

Cinelerra's source code is available under the GNU General Public
License (GPL).  However, unlike most large Free Open Source projects,
the development of Cinelerra is not open to distributed collaboration
and there is no support for the software.

Cinelerra is tested on Fedora. Cinelerra4 is available also as Ubuntu
packages.

* About Cinelerra-CV

CV stands for Community Version. This website is meant to be a
community home page for Cinelerra.  We try to complete the amazing
work of Heroine Virtual Ltd. (HV) offering also a place for
collaborative development and community help.