Import php5-pear-XML_Parser-1.3.2 as wip/pear-XML_Parser.
pear-XML_Parser is an XML parser based on PHPs built-in xml extension.
It supports two basic modes of operation: "func" and "event". In "func"
mode, it will look for a function named after each element (xmltag_ELEMENT
for start tags and xmltag_ELEMENT_ for end tags), and in "event" mode it
uses a set of generic callbacks.