Import pan2-git-2nb20100228 as wip/pan2-git.
Pan is a Newsreader for Gnome whose goal is to be easy to use and
functional enough for beginners and advanced users alike.  It has the
basic features of threading, reading, filtering, posting, and offline
reading, and also has a number of features for power users and
alt.binaries fans.

It's the only Unix newsreader to get a perfect score on the Good
Net-Keeping Seal of Approval evalutions.

This is the bug-fixing version of Pan maintained by K. Haley.