Import eric4-4.4.1nb1 as wip/eric4.
Eric is a full featured Python and Ruby editor and IDE, written in python.
It is based on the cross platform Qt gui toolkit, integrating the highly
flexible Scintilla editor control. It is designed to be usable as everdays'
quick and dirty editor as well as being usable as a professional project
management tool integrating many advanced features Python offers the
professional coder. eric4 includes a plugin system, which allows easy
extension of the IDE functionality with plugins downloadable from the net.