Import tDOM-0.8.2 as wip/tcl-tDOM.
tDOM combines high performance XML data processing with easy and
powerful Tcl scripting functionality. tDOM should be one of the
fastest ways to manipulate XML with a scripting language and uses very
little memory in the process (for example the DOM tree of the XML
recommendation in XML (160K) needs only about 450K in memory)! The
tDOM project was started by Jochen Loewer. It is currently maintained
by Rolf Ade.